+++ krcg/seating.py
@@ -3,11 +3,13 @@
 # When organising a tournament, it's important to devise an equitable seating
 # which also ensures a maximum of diversity over rounds and tables.
 
-from typing import Callable, Iterable, List, Tuple
+from typing import Callable, Dict, Hashable, Iterable, List, Tuple
 import collections
+import concurrent.futures
 import math
 import numpy
 import itertools
+import multiprocessing
 import random
 
 # The seating rules: code, label, weight
@@ -39,19 +41,90 @@ POSITION_MEASURE = {
 
 
 class Round(list):
-    def __init__(self, permutation: List[int]):
-        length = len(permutation)
+    """A list of list representing the tables of a round"""
+
+    @classmethod
+    def from_players(cls, players: Iterable[Hashable]):
+        """Build a round structure from a simple list of players"""
+        length = len(players)
+        if length in {6, 7, 11}:
+            raise ValueError(
+                f"A staggered round structure is required for {length} players"
+            )
         fours = 5 - (length % 5 or 5)
         fives = (length - 4 * fours) // 5
-        super().__init__(
+        return cls(
             itertools.chain(
-                (permutation[i : i + 5] for i in range(0, fives * 5, 5)),
+                (players[i : i + 5] for i in range(0, fives * 5, 5)),
                 (
-                    permutation[i : i + 4]
+                    players[i : i + 4]
                     for i in range(fives * 5, fives * 5 + fours * 4, 4)
                 ),
             )
         )
+
+    @classmethod
+    def copy(cls, round_):
+        """Copy another Round"""
+        return cls(t[:] for t in round_.iter_tables())
+
+    def shuffle(self):
+        players = list(self.iter_players())
+        random.shuffle(players)
+        for i in range(self.players_count()):
+            self[i] = players[i]
+
+    def iter_table_players(self) -> Iterable[Tuple[int, int, int, Hashable]]:
+        """Full information players iteration
+
+        Yields: table number, position, table size, player number
+        """
+        for table_number, players in enumerate(self, 1):
+            table_size = len(players)
+            for position, player in enumerate(players, 1):
+                yield table_number, position, table_size, player
+
+    def iter_tables(self):
+        """Iterate on the tables, not the players (default)"""
+        yield from super().__iter__()
+
+    def iter_players(self):
+        """Convenience method for symmetry with `iter_tables`"""
+        for table in super().__iter__():
+            for player in table:
+                yield player
+
+    def tables_count(self):
+        """Get the number of tables (default len gives the number of players)"""
+        return super().__len__()
+
+    def players_count(self):
+        """Convenience function for symmetry with `tables_count`"""
+        return sum(len(table) for table in self.iter_tables())
+
+    def __global_index_to_tuple(self, index: int):
+        """Private method to get the underlying index out of a naive global index"""
+        table_index = 0
+        for table in self.iter_tables():
+            if index >= len(table):
+                index -= len(table)
+                table_index += 1
+                continue
+            return table_index, index
+        else:
+            raise IndexError("Out of bounds")
+
+    def __getitem__(self, index: int):
+        """Access items transparently using a global index"""
+        i, j = self.__global_index_to_tuple(index)
+        return super().__getitem__(i)[j]
+
+    def __setitem__(self, index, value: Hashable):
+        """Access items transparently using a global index"""
+        i, j = self.__global_index_to_tuple(index)
+        super().__getitem__(i)[j] = value
+
+    # NOTE: Do not mess with default iteration to avoid problems with multiprocessing
 
 
 Measure = collections.namedtuple("Measure", ["position", "opponents"])
@@ -61,9 +134,9 @@ Measure.__radd__ = lambda rhs, lhs: rhs if lhs == 0 else rhs.__add__(lhs)
 
 def measure(max_player_number: int, round_: Round) -> Measure:
     """Measure a round (list of tables), returns two matrices:
-    position (players_count x 7):
+    position (players_count x 8):
         for each player:
-            vps, transfers (integer),
+            played, vps, transfers (integer),
             seat1, seat2, seat3, seat4, seat5 (1 for the seat they occupy)
     opponents (players_count x players_count x 8):
         for each pair of players, booleans indicating if they were:
@@ -74,14 +147,16 @@ def measure(max_player_number: int, round_: Round) -> Measure:
     This allows to re-compute a single round measure when a single round is changed.
     max_player_number must be the max over all rounds so that we can add round matrices.
     """
-    position = numpy.zeros((max_player_number, 7), int)
+    position = numpy.zeros((max_player_number, 8), int)
     opponents = numpy.zeros((max_player_number, max_player_number, 8), int)
-    for table in round_:
+    for table in round_.iter_tables():
         table_size = len(table)
         for seat, player in enumerate(table):
-            position[player - 1][0] = table_size
-            position[player - 1][1] = min(seat + 1, 4)
-            position[player - 1][2 + seat] = 1
+            position[player - 1][0] = 1
+            position[player - 1][1] = table_size
+            transfers = min(seat + 1, 4)
+            position[player - 1][2] = transfers
+            position[player - 1][3 + seat] = 1
             for relation, opponent in enumerate(
                 itertools.chain(table[seat + 1 :], table[:seat])
             ):
@@ -105,7 +180,7 @@ Deviation = collections.namedtuple("Deviation", ["player", "value"])
 class Score:
     """A detailed scoring of a seating measure.
 
-    Original criteria
+    Original criteria:
     https://groups.google.com/g/rec.games.trading-cards.jyhad/c/4YivYLDVYQc/m/CCH-ZBU5UiUJ
 
     R1 No pair of players repeat their predator-prey relationship. This is mandatory.
@@ -151,24 +226,34 @@ class Score:
         )
     )
 
-    def __init__(self, measure: Measure, mask=None):
+    def __init__(self, rounds: List[Round], max_player_number: int = None):
+        max_player_number = max_player_number or _max_player_number(rounds)
+        self.score_measure(
+            sum(measure(max_player_number, r) for r in rounds), len(rounds)
+        )
+
+    def score_measure(self, measure: Measure, rounds_count: int) -> None:
         # transfers, starting vps: compute standard deviation
-        masked = Score._get_masked(measure, mask)
-        self.R3, self.R8 = numpy.std(masked, 0)
+        vps = measure.position[:, 1] / measure.position[:, 0]
+        transfers = measure.position[:, 2] / measure.position[:, 0]
+        self.R3 = numpy.std(vps)
+        self.R8 = numpy.std(transfers)
         # record details of anomalies for output
-        self.mean_vps, self.mean_transfers = numpy.mean(masked, 0)
-        vps, transfers = numpy.transpose(masked)
+        self.mean_vps = numpy.mean(vps)
+        self.mean_transfers = numpy.mean(transfers)
         self.vps = [
             Deviation(i + 1, vps[i])
-            for i in numpy.flatnonzero(abs(self.mean_vps - vps) > 0.5)
+            for i in numpy.flatnonzero(abs(self.mean_vps - vps) > 1 / rounds_count)
         ]
         self.transfers = [
             Deviation(i + 1, transfers[i])
-            for i in numpy.flatnonzero(abs(self.mean_transfers - transfers) > 0.5)
+            for i in numpy.flatnonzero(
+                abs(self.mean_transfers - transfers) > 1 / rounds_count
+            )
         ]
         # same seat twice (or more)
         self.R7 = [
-            SeatViolation(*v) for v in numpy.argwhere(measure.position[:, 2:] > 1) + 1
+            SeatViolation(*v) for v in (numpy.argwhere(measure.position[:, 3:] > 1) + 1)
         ]
         # fifth seat twice (or more)
         self.R5 = [PlayerViolation(a) for a, s in self.R7 if s == 5]
@@ -207,51 +292,42 @@ class Score:
         self.total = sum(x * m for x, m in zip(self.rules, [R[2] for R in RULES]))
 
     @staticmethod
-    def fast_total(measure: Measure, mask=None):
+    def fast_total(measure: Measure) -> float:
         """Get just a total score of a seating measure (all rounds).
 
         This is used to speed up computations when searching for an optimum.
         """
-        masked = Score._get_masked(measure, mask)
         rules = [
             len(numpy.argwhere(measure.opponents[:, :, 1] > 1)),
             len(numpy.argwhere(measure.opponents[:, :, 0] > 2)) // 2,
-            numpy.std(masked[:, 0], 0),
+            numpy.std(measure.position[:, 1] / measure.position[:, 0]),
             len(numpy.argwhere(measure.opponents[:, :, 0] > 1)) // 2,
             len(numpy.argwhere(measure.position[:, 6] > 1)),
             len(numpy.argwhere(measure.opponents[:, :, 1:6] > 1)) // 2,
-            len(numpy.argwhere(measure.position[:, 2:] > 1)),
-            numpy.std(masked[:, 1], 0),
+            len(numpy.argwhere(measure.position[:, 3:] > 1)),
+            numpy.std(measure.position[:, 2] / measure.position[:, 0]),
             len(numpy.argwhere(measure.opponents[:, :, 6:] > 1)) // 2,
         ]
         return sum(x * m for x, m in zip(rules, [R[2] for R in RULES]))
 
-    @staticmethod
-    def _get_masked(measure, add_mask):
-        masked = measure.position[:, :2]
-        mask = masked == 0
-        if add_mask is not None:
-            mask |= add_mask
-        return numpy.ma.MaskedArray(masked, mask)
 
-
-def permutations(players_count: int, rounds_count: int):
-    """Return the base permutations for given parameters
+def get_rounds(players_count: int, rounds_count: int) -> List[Round]:
+    """Return the base rounds for given parameters
 
     This gets complicated only if you got 6, 7 or 11 players, otherwise it's simply
     the list of all players for each round.
 
-    >>> permutations(8, 3)
-    [[1, 2, 3, 4, 5, 6, 7, 8],
-     [1, 2, 3, 4, 5, 6, 7, 8],
-     [1, 2, 3, 4, 5, 6, 7, 8]]
+    >>> rounds(8, 3)
+    [[[1, 2, 3, 4], [5, 6, 7, 8]],
+     [[1, 2, 3, 4], [5, 6, 7, 8]],
+     [[1, 2, 3, 4], [5, 6, 7, 8]]]
     """
     if players_count < 4:
         raise RuntimeError("At least 4 players required")
 
     if players_count not in [6, 7, 11]:
         base = list(range(1, players_count + 1))
-        return [base[:] for _ in range(rounds_count)]
+        return [Round.from_players(base[:]) for _ in range(rounds_count)]
 
     if rounds_count < 2:
         raise RuntimeError("At least 2 rounds by player are required")
@@ -299,35 +375,32 @@ def permutations(players_count: int, rounds_count: int):
         excludes -= possible_outs[i]
     exclusions = sum([[i] * out[i] for i in range(rounds_count)], [])
     return [
-        [
-            p
-            for p in range(1, players_count + 1)
-            if not any(
-                exclusions[p - 1 + players_count * c] == r
-                for c in range(additional_rounds)
-            )
-        ]
+        Round.from_players(
+            [
+                p
+                for p in range(1, players_count + 1)
+                if not any(
+                    exclusions[p - 1 + players_count * c] == r
+                    for c in range(additional_rounds)
+                )
+            ]
+        )
         for r in range(rounds_count)
     ]
 
 
-def score_rounds(rounds: List[Round]) -> Measure:
-    """Provide the score for a full seating
-
-    Convenience function
-    """
-    max_player_number = max(
-        p for p in itertools.chain.from_iterable(t for r in rounds for t in r)
+def _max_player_number(rounds: List[Round]) -> int:
+    """Internal function used to get the matrix dimension required for measures."""
+    return max(
+        p for p in itertools.chain.from_iterable(r.iter_players() for r in rounds)
     )
-    return Score(sum(measure(max_player_number, r) for r in rounds))
 
 
 def optimise(
-    permutations: list,
+    rounds: List[Round],
     iterations: int,
+    fixed: int = None,
     callback: Callable = None,
-    fixed: int = 1,
-    ignore: Iterable[int] = None,
 ) -> Tuple[List[Round], Score]:
     """Given a list of players for each round, compute an optimal seating.
 
@@ -338,62 +411,62 @@ def optimise(
         * trials (since last callback call)
         * accepts (since last callback call)
         * improves (since last callback call)
-    - fixed is the number of permutations that are left untouched by the optimisation
-    - ignore is a list of players numbers to ignore when calculating
-      VPs and transfers deviation: players not playing all rounds should be listed
+    - fixed is the number of rounds that are left untouched by the optimisation
+      (default is all except the last one)
 
     Use a simulated annealing algorithm:
         - exponential cooldown strategy
         - given the problem shape, reset the state to the best known one regularily
 
-    Using a list of players per round as entry allows the function to be used for
+    Using a list of tables per round as entry allows the function to be used for
     corner cases like:
         - 6, 7 and 11 players seatings
         - change of players in the middle of a tournament
+
+    Iterations:
+        empyrism shows that the best results are achieved calling
+        this 4x with around 20k iterations each for a full 3R+F round structure
+
+        For successive calls to build each round as they come,
+        players going in and out of them, a simple 20k iteration is sufficient.
     """
     random.seed()
     # annealing parameters have been chosen experimentally
-    # verbose is kept here to be set to True in case there's a need to re-examine them
     temperature_min = 0.001
     temperature_max = RULES[0][2]
     temperature_factor = -math.log(temperature_max / temperature_min)
-    max_player_number = max(p for p in itertools.chain.from_iterable(permutations))
-    mask = None
-    if ignore:
-        mask = numpy.zeros((max_player_number, 2), bool)
-        for player in ignore:
-            mask[player - 1, :2] = 1
+    max_player_number = _max_player_number(rounds)
     # initial state
-    for permutation in permutations[fixed:]:
-        random.shuffle(permutation)
+    if fixed is None:
+        fixed = len(rounds) - 1
     temperature = temperature_max
-    measures = [
-        measure(max_player_number, Round(permutation)) for permutation in permutations
-    ]
-    best_score = previous_score = score = Score.fast_total(sum(measures), mask)
-    best_state = [p[:] for p in permutations]
+    measures = [measure(max_player_number, r) for r in rounds]
+    best_score = previous_score = score = Score.fast_total(sum(measures))
+    best_state = [Round.copy(r) for r in rounds]
+    for round_ in rounds[fixed:]:
+        round_.shuffle()
     trials, accepts, improves = 0, 0, 0
 
     # Exploration
     for step in range(iterations):
         temperature = temperature_max * math.exp(temperature_factor * step / iterations)
-        round_index = random.randrange(fixed, len(permutations))
-        permutation = permutations[round_index]
-        # note all permutations may not have the same length (eg. 6, 7, 11 players)
-        length = len(permutation)
+        round_index = random.randrange(fixed, len(rounds))
+        round_ = rounds[round_index]
+        # note all rounds might not have the same length
+        length = round_.players_count()
         i = random.randrange(length)
         j = random.randrange(length)
-        permutation[i], permutation[j] = permutation[j], permutation[i]
+        round_[i], round_[j] = round_[j], round_[i]
         # only recompute the changed round, other rounds have not varied
         previous_measure = measures[round_index]
-        measures[round_index] = measure(max_player_number, Round(permutation))
-        score = Score.fast_total(sum(measures), mask)
+        measures[round_index] = measure(max_player_number, round_)
+        score = Score.fast_total(sum(measures))
         score_diff = score - previous_score
         trials += 1
         # accept or reject the move depending on its score and temperature
         # the higher temperature, the higher the chance to accept a non-improving move
         if score_diff > 0 and math.exp(-score_diff / temperature) < random.random():
-            permutation[i], permutation[j] = permutation[j], permutation[i]
+            round_[i], round_[j] = round_[j], round_[i]
             score = previous_score
             measures[round_index] = previous_measure
         else:
@@ -402,7 +475,7 @@ def optimise(
             if score_diff < 0.0:
                 improves += 1
             if score < best_score:
-                best_state = [p[:] for p in permutations]
+                best_state = [Round.copy(r) for r in rounds]
                 best_score = score
         # every 100th of the way, call the callback
         # and reset the state to the best known state
@@ -417,11 +490,23 @@ def optimise(
                     improves=improves,
                 )
             trials, accepts, improves = 0, 0, 0
-            permutations = [p[:] for p in best_state]
-            measures = [
-                measure(max_player_number, Round(permutation))
-                for permutation in permutations
-            ]
+            rounds = [Round.copy(r) for r in best_state]
+            measures = [measure(max_player_number, r) for r in rounds]
             previous_score = best_score
 
-    return [Round(p) for p in permutations], Score(sum(measures), mask)
+    return best_state, Score(best_state, max_player_number=max_player_number)
+
+
+def archon_seating(players_count: int, rounds_per_player: int):
+    """Convenience function to compute a full multiround seating"""
+    rounds = get_rounds(players_count, rounds_per_player)
+    try:
+        cpus = multiprocessing.cpu_count()
+    except NotImplementedError:
+        cpus = 1
+    results = []
+    with concurrent.futures.ProcessPoolExecutor() as executor:
+        for _ in range(cpus):
+            results.append(executor.submit(optimise, rounds, 80000, 1))
+        results = [r.result() for r in results]
+        return min(results, key=lambda x: x[1].total)